feat(musl): add musl toolchain (#2402)
With this and the previously landed #2406 users can use this in `sdist`
or `whl`-only setups alike. The PR itself is very simple and just adds
`musl` toolchains.

Whilst at it also move the `WhlLibc` flag out of the `pypi` namespace
since it is also used for `toolchain` matching now.

Fixes #1211