fix: make sys.executable work with script bootstrap (#2409)
When `--bootstrap_impl=script` is used, `PYTHONPATH` is no longer used
to set the import
paths, which means subprocesses no longer inherit the Bazel paths. This
is generally a
good thing, but breaks when `sys.executable` is used to directly invoke
the interpreter.
Such an invocation assumes the interpreter will have the same packages
available and works
with the system_python bootstrap.

To fix, have the script bootstrap use a basic virtual env. This allows
it to intercept
interpreter startup even when the Bazel executable isn't invoked. Under
the hood, there's
two pieces to make this work. The first is a binary uses
`declare_symlink()` to write a
relative-path based symlink that points to the underlying Python
interpreter. The second
piece is a site init hook (triggered by a `.pth` file using an `import`
line) performs
sys.path setup as part of site (`import site`) initialization.

Fixes https://github.com/bazelbuild/rules_python/issues/2169

---------

Co-authored-by: Ignas Anikevicius <240938+aignas@users.noreply.github.com>