fix: normalize argv0 so runfiles root can be found on windows with bazel 9 (#2481)
When the shell test invokes the python binary, it uses a combination of
forward slashes and
backslashes. Under Bazel 9, that mixture of slashes is preserved. This
later breaks a regex
that looks for the OS-specific path separator.

To fix, normalize forward slashes to the OS path separator.

Oddly, it's not Bazel that is passing the mixture of slashes (it's the
shell), but behavior seems to
vary based on which version of Bazel is used.

Along the way, copy the nicer `print_verbose` function from the stage2
bootstrap into the old
bootstrap. It prints debug information in a nicer format.

Work towards https://github.com/bazelbuild/rules_python/issues/2469