chore: remove find_requirements re-export from bazel_tools (#2484)
This removes re-exporting the `find_requirements` aspect from
`@bazel_tools`.

The find_requirements aspect is only useful for analyzing the
`srcs_version` relationship
between targets, which was only needed as part of the Python 2 to 3
upgrade. With Python
2 no longer supported, it's defunct.

Removing it entirely because I can't find any real references to it in
the wild. It looks
entirely unused.