fix(pip.parse): pass requirements without env markers to the whl_library
With this change the environment markers from the requirements.txt files
no longer end up in the whl_library definitions. The alternative would be
to fix this in the whl_library itself and do the string manipulation then.
However, this means that we will be doing refetching of the repository when
the markers change and the overall behaviour may be more complex. This
solution also makes the MODULE.bazel.lock files simpler.

That said, I am 50/50 on putting it in this way, so I can be easily convinced
to put it in the whl_library if there is preference.