fix(pypi): Fix use_hub_alias_dependencies with WORKSPACE
The code path pip_parse follows when using a WORKSPACE file with
use_hub_alias_dependencies enabled forgets to pass requirement cycles
along to alias creation, leading to the _groups package never being
created and aliases skipping them.  Requirement cycles are just ignored
entirely.  In this patch we attempt to fix that so grouping works more
or less the same way as it does under bzlmod with that flag enabled.