fix(pypi): change the parallelisation scheme for querying SimpleAPI
Instead of querying everything in parallel and yielding a lot of 404
warnings, let's query the main index first and then query the other
indexes only for the packages that were not yet found. This should make
the problem in #2100 not as serious.

What is more, we can print the value of  for the users
to use.

Work towards #2100