fix(gazelle): Don't ignore `setup.py` files when running Gazelle (#2536)
Don't ignore `setup.py` files when running Gazelle.

Fixes #2108.

I believe that `setup.py` was originally ignored because it, when found
that the repo root, is part of `setuptools` config and may have caused
problems with Gazelle.

I've been running our Google Quantum code with this patch for a long
while now and not seen any issues. I figured it was time to upstream it.