feat: Add feature to expose whether the native rules are used (#2549)
Internally rules_python decides whether the implementation should use
the legacy builtin rules from bazel or not.
The rules_python attributes have diverged from the builtin rules e.g.
`precompile` and so a consumer of this library that wants to support
bazel versions that are still using the legacy builtin rules needs a
method to be able to query whether the rules are being used.

This change adds a entry to features.bzl to expose whether the legacy
builtin rules are being used.