chore: remove internal usage of deprecated py_binary ad py_test (#2569)
This goes together with #2565 to remove the internal usage of the
deprecated
symbols. This also fixes the compile_pip_requirements symbol to print
the
correct deprecation message.

Builds on top of 611eda8

The example message that would be printed is as follows:
```
The 'py_test' symbol in '@+python+python_3_11//:defs.bzl'
is deprecated. It is an alias to the regular rule; use it directly instead:

load("@rules_python//python:py_test.bzl", "py_test")

py_test(
    name = "versioned_py_test",
    srcs = ["dummy.py"],
    main = "dummy.py",
    python_version = "3.11.11",
)
```

---------

Co-authored-by: Richard Levasseur <richardlev@gmail.com>