fix: make plain zipapp work with bootstrap=script (#2598)
The `__main__.py` template (zip_main_template.py) was using the wrong
path when creating
the interpreter symlinks. It as computing it correctly, just the wrong
variable was used
in the symlink() call.

To fix, pass the correct variable.

Also adds a test to check that it's runnable.

Fixes https://github.com/bazelbuild/rules_python/issues/2596