refactor: expose base rule construction via builders to allow customization for testing (#2600)
The py_reconfig rules work by wrapping: The outer reconfig rule applies
a transition, depends
on an inner py base rule, then jumps through various hoops to ensure it
looks and acts
like the target it's wrapping.

This is error prone, incomplete, and annoying code to maintain. Phil
recently discovered it wasn't properly propagating the output group, so
he had to add that. I wasted time trying to fix a bug I _thought_ was in
it, but actually was working correctly. The logic within it is a bit
hacky as it tries to emulate some of the platform-specific stuff for
windows. Every time
py_reconfig gains something to transition on, there's numerous places to
define, propagate,
and extract the pieces necessary to do it.

To fix this, make the py_reconfig rules not wrap an inner base py rule.
Instead, they use the same underlying rule args that the base rules do.
This lets them act directly as the rule they're designed to test.

Customization is done by capturing all the rule args in builder objects.
The py_reconfig code constructs the same builder the base rules do, then
modifies it as necessary (adding attributes, wrapping the base
transition function). As a bonus, this sets some ground work to allow
more easily defining derivative rules without having to copy/paste
arbitrary parts of how the base rules are defined.

Work towards https://github.com/bazelbuild/rules_python/issues/1647