fix: Gazelle bug with merging py_binary targets in per-file mode and partial update (#2619)
This PR adds a new unit test. Currently, this is just a failing test
without a fix, and I am still trying to understand the code well enough
to find the root cause of the issue.

Our team uses Python+Gazelle in a monorepo, and we have a handful of
directories with multiple `.py` files containing `if __name__ ==
"__main__"`. Most of the time these are present for convenience or
ad-hoc invocation. We're aware of the [recommendation to split these
into separate
files](https://github.com/bazelbuild/rules_python/tree/main/gazelle#binaries),
but that can cause clutter, and it is non-obvious to most engineers what
to do when encountering this issue, which presents either as a
misleading error message or a no-op without creating the appropriate
targets.

**Update**
This bug occurs when ALL of the following are true:
* `python_generation_mode` is set to `file`.
* Multiple python binary files (files with `if __name__ == "__main__"`)
exist in the same directory.
* The directory has no `__main__.py` file.
* The `BUILD` file in the directory is partially complete, i.e. it
contains `py_binary` targets for some of the python files, but not
others.

In this situation, previously absent `py_binary` targets are merged into
existing `py_binary` targets instead of being created as new targets.

---------

Co-authored-by: Jimmy Tanner <jimmy@physicalintelligence.company>