refactor: stop using some deprecated Starlark APIs (#2626)
I am currently working on an analysis tool that is capable of parsing
BUILD/*.bzl files. It currently fails to process some of the Python
rules, due to the rules depending on some features that are deprecated
on the Bazel side. Instead of adding implementations of these deprecated
features to my brand new analysis tool, I thought I'd simply patch up
the Python rules instead.