feat: allow merging multiple simpleapi indexes
Before this change users would have no way to download pytorch from
multiple indexes (e.g. `cpu` version on some platforms and `gpu` on
another) and use it inside a single hub repository.

This brings the users necessary toggles to tell `rules_python` to search
in multiple indexes for the `torch` wheels. Note, that the
`index_strategy` field is going to be used only if one is setting
multiple indexes via the `extra_index_urls` and not via the
`index_url_overrides`.

Whilst at it I have improved the `simpleapi_download` tests to also test
the warning messages that we may print to the user.

Fixes #2622