feat: uv lock rule instead of genrule
This change implements the uv pip compile as a rule.

In order to also make things easier to debug we provide
a runnable rule that has the same arguments and updates
the source tree output file automatically.

The main design is to have a regular lock rule and then
it returns a custom provider that has all of the recipe
ingredients to construct an executable rule. The execution
depends on having bash or powershell, however the
powershell script is not yet complete and requires some
help from the community.

Work towards #1975.

Address all of the comments