feat(rules): allow deriving custom rules from core rules (#2666)
This exposes public functions for creating builders for py_binary,
py_test, and py_library.
It also adds some docs and examples for how to use them.

I'm calling this a "volatile" API -- it's public, but the pieces that
comprise
it (e.g. all the rule args, attributes, the attribute args, etc) are
likely to change
in various ways, and not all modifications to them can be supported in a
backward
compatible way. Hence the "volatile" term:
* hold it gently and its fine
* shake it a bit and its probably fine
* shake it moderately and something may or may not blow up
* shake it a lot and something will certainly blow up.

Work towards https://github.com/bazelbuild/rules_python/issues/1647

---------

Co-authored-by: Ignas Anikevicius <240938+aignas@users.noreply.github.com>