feat(rules): add main_module attribute to run a module name (python -m) (#2671)
This implements the ability to run a module name instead of a file path,
aka
`python -m` style of invocation.

This allows a binary/test to specify what the main module is without
having to
have a direct dependency on the entry point file.

As a side effect, the `srcs` attribute is no longer required.

Fixes https://github.com/bazelbuild/rules_python/issues/2539