Explicitly watch the python version file we read
If we read a python_version_file (specified with the defaults tag
class or the default @@//:.python-version file), explicitly start
watching that file. There are some circumstances where such a watch
wouldn't be allowed, but since it's new functionality it doesn't break
anything for anyone to insist on the watch. If a specific use case
requires the relaxation of the requirement, that can always be
considered later.