chore: remove old versions of Python 3.8 (#2700)
Python 3.8 has reached EOL and this PR removes old toolchains and most
of the tests. Users can still use it if they register the toolchains
themselves, but `rules_python` will no longer keep testing the
toolchains.

Removing the toolchains all-together will be done at a later stage which
may require us to be more clever how we handle asks to include `3.8`.
Maybe
we can just fail if the user asks for a python version that does not
exist,
but I am concerned that `rules_python` depending on `protobuf` may pull
in
code that requests `3.8`. I'll look at this at some later time.