Stop attempting to read @@//:.python-version
The idea to read .python-version in the root module was problematic
for two reasons. For one thing, if there is a .python-version file in
the root module, but no BUILD or BUILD.bazel file on the top level, we
produce an error that we can't catch. This would be a breaking change,
requiring users to add a BUILD.bazel file. Also, if the user has a
.python-version file, and also sets is_default on a toolchain, we
would fail the build (or, if we didn't, we'd change the default Python
version against the users explicit instructions). That would also be a
breaking change.

Removing the @@//:.python-version logic means the user needs to opt in
by saying python_version_file="@@//.python-version". That's not
necessarily a bad thing, as being explicit allows anyone to grep the
repo for the label and find out what cares about that file (as pointed
out by @fmeum on Slack; good point).