feat(python.toolchain): support file-based default Python version (#2588)
This change adds a new `default_version_file` attribute to
`python.toolchain`. If set, the toolchain compares the file's contents
to its `python_version`, and if they match, treats that toolchain as
default (ignoring `is_default`). This allows Bazel to synchronize the
default Python version with external tools (e.g., pyenv) that use a
`.python-version` file or environment variables.

Fixes #2587.

---------

Co-authored-by: Ignas Anikevicius <240938+aignas@users.noreply.github.com>